+++ src/app/Balloon.App.Notification/Notifier.php
@@ -20,6 +20,7 @@ use MongoDB\BSON\ObjectId;
 use MongoDB\BSON\UTCDateTime;
 use MongoDB\Database;
 use Psr\Log\LoggerInterface;
+use InvalidArgumentException;
 
 class Notifier
 {
@@ -73,14 +74,49 @@ class Notifier
     protected $template;
 
     /**
+     * Notification throttle.
+     *
+     * @var int
+     */
+    protected $notification_throttle = 120;
+
+    /**
      * Constructor.
      */
-    public function __construct(Database $db, Server $server, LoggerInterface $logger, TemplateHandler $template)
+    public function __construct(Database $db, Server $server, LoggerInterface $logger, TemplateHandler $template, array $config=[])
     {
         $this->logger = $logger;
         $this->db = $db;
         $this->server = $server;
         $this->template = $template;
+        $this->setOptions($config);
+    }
+
+    /**
+     * Set config.
+     */
+    public function setOptions(array $config = []): self
+    {
+        foreach ($config as $option => $value) {
+            switch ($option) {
+                case 'notification_throttle':
+                    $this->{$option} = (int) $value;
+
+                break;
+                default:
+                    throw new InvalidArgumentException('invalid option '.$option.' given');
+            }
+        }
+
+        return $this;
+    }
+
+    /**
+     * Get notification throttle time
+     */
+    public function getThrottleTime(): int
+    {
+        return $this->notification_throttle;
     }
 
     /**
@@ -265,13 +301,11 @@ class Notifier
     /**
      * Throttle subscriptions.
      */
-    public function throttleSubscriptions(NodeInterface $node, array $user): Notifier
+    public function throttleSubscriptions(array $subscriptions): Notifier
     {
-        $node_id = $node->isReference() ? $node->getShareId() : $node->getId();
         $this->db->subscription->updateMany([
-            'node' => $node_id,
-            'user' => [
-                '$in' => $user,
+            '_id' => [
+                '$in' => $subscriptions,
             ],
         ], [
             '$set' => [
@@ -324,6 +358,12 @@ class Notifier
             }
         }
 
+        $parents = [$ids[0]];
+
+        if(count($ids) > 1) {
+            $parents[] = $ids[1];
+        }
+
         return $this->db->subscription->find([
             '$or' => [
                 [
@@ -331,7 +371,7 @@ class Notifier
                     'recursive' => true,
                 ],
                 [
-                    'node' => ['$in' => [$ids[0],$ids[1]]],
+                    'node' => ['$in' => $parents],
                 ]
             ]
         ]);
@@ -340,7 +380,7 @@ class Notifier
     /**
      * Subscribe to node updates.
      */
-    public function subscribeNode(NodeInterface $node, bool $subscribe = true, bool $exclude_me = true, bool $recursive = false): bool
+    public function subscribeNode(NodeInterface $node, bool $subscribe = true, bool $exclude_me = true, bool $recursive = false, ?int $throttle=null): bool
     {
         $node_id = $node->isReference() ? $node->getShareId() : $node->getId();
         $user_id = $this->server->getIdentity()->getId();
@@ -354,6 +394,7 @@ class Notifier
                 'timestamp' => new UTCDateTime(),
                 'exclude_me' => $exclude_me,
                 'recursive' => $recursive,
+                'throttle' => $throttle,
                 'user' => $user_id,
                 'node' => $node_id,
             ];
@@ -368,23 +409,6 @@ class Notifier
                 'upsert' => true,
             ]
             );
-
-            /*if ($node instanceof Collection && $recursive === true) {
-                $db = $this->db;
-                $node->doRecursiveAction(function ($child) use ($db, $subscription) {
-                    $subscription['node'] = $child->getId();
-                    $db->subscription->replaceOne(
-                        [
-                        'user' => $subscription['user'],
-                        'node' => $subscription['node'],
-                    ],
-                        $subscription,
-                        [
-                        'upsert' => true,
-                    ]
-                    );
-                });
-            }*/
         } else {
             $this->logger->debug('user ['.$this->server->getIdentity()->getId().'] unsubscribes node ['.$node->getId().']', [
                 'category' => get_class($this),